Switch to int and cover reward with edge case (partial covering by multiple inventories)

+++ core/model/balance.go
@@ -3,5 +3,5 @@ package model
 // RewardTypeAmount wraps the balance aggregation response
 type RewardTypeAmount struct {
 	RewardType string `json:"reward_type" bson:"_id"`
-	Amount     int64  `json:"amount" bson:"amount"`
+	Amount     int    `json:"amount" bson:"amount"`
 } // @name RewardTypeAmount
